+++ packages/plugin-javascript/CHANGELOG.md
@@ -9,6 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Added `typescript` option when calling `javascript()` which when set to true will add typescript babel presets. This replaces plugin-typescript's `typescript` project plugin.
 - Removed js output from `build/ts` folder. [[#217](https://github.com/Shopify/sewing-kit-next/pull/217)]
 
 ## 0.6.0 - 2021-07-07